+++ project/Settings.scala
@@ -20,21 +20,16 @@ object Settings {
     "-feature"
   )
 
-  /** Set some basic options when running the project with Revolver */
-  val jvmRuntimeOptions = Seq(
-    "-Xmx1G"
-  )
-
   /** Declare global dependency versions here to avoid mismatches in multi part dependencies */
   object versions {
     val scala = "2.11.7"
     val scalaDom = "0.8.2"
-    val scalajsReact = "0.10.2-SNAPSHOT"
+    val scalajsReact = "0.10.2"
     val scalaCSS = "0.3.1"
     val log4js = "1.4.10"
     val autowire = "0.2.5"
     val booPickle = "1.1.0"
-    val diode = "0.1.0-SNAPSHOT"
+    val diode = "0.1.0"
     val uTest = "0.3.1"
 
     val react = "0.14.3"
@@ -74,8 +69,8 @@ object Settings {
 
   /** Dependencies for external JS libs that are bundled into a single .js file according to dependency order */
   val jsDependencies = Def.setting(Seq(
-    "org.webjars.bower" % "react" % "0.14.3" / "react-with-addons.js" minified "react-with-addons.min.js" commonJSName "React",
-    "org.webjars.bower" % "react" % "0.14.3" / "react-dom.js" minified "react-dom.min.js" dependsOn "react-with-addons.js" commonJSName "ReactDOM",
+    "org.webjars.bower" % "react" % versions.react / "react-with-addons.js" minified "react-with-addons.min.js" commonJSName "React",
+    "org.webjars.bower" % "react" % versions.react / "react-dom.js" minified "react-dom.min.js" dependsOn "react-with-addons.js" commonJSName "ReactDOM",
     "org.webjars" % "jquery" % versions.jQuery / "jquery.js" minified "jquery.min.js",
     "org.webjars" % "bootstrap" % versions.bootstrap / "bootstrap.js" minified "bootstrap.min.js" dependsOn "jquery.js",
     "org.webjars" % "chartjs" % versions.chartjs / "Chart.js" minified "Chart.min.js",